refactor(userinterface): userinterface 분리
- 기존의 InputHandler.java를 InputValidator, UserRequest, UserResponse로 분리
- User의 input과 output을 처리하는 클래스 UserRequest, UserResponse 생성
- User와 상호작용하는 클래스 userinterface로 패키징
